+++ lib/src/commit_builder.rs
@@ -66,6 +66,14 @@ impl CommitBuilder {
         let mut commit = predecessor.store_commit().clone();
         commit.predecessors = vec![predecessor.id().clone()];
         commit.committer = settings.signature();
+        // If the user had not configured a name and email before but now they have,
+        // update the the author fields with the new information.
+        if commit.author.name == UserSettings::user_name_placeholder() {
+            commit.author.name = commit.committer.name.clone();
+        }
+        if commit.author.email == UserSettings::user_email_placeholder() {
+            commit.author.email = commit.committer.email.clone();
+        }
         CommitBuilder {
             store: store.clone(),
             commit,