+++ CHANGELOG.md
@@ -99,6 +99,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * `jj log` now accepts a `--reversed` option, which will show older commits
   first.
 
+* The "(no name/email configured)" placeholder value for name/email will now be
+  replaced if once you modify a commit after having configured your name/email.
+
 ### Fixed bugs
 
 * When rebasing a conflict where one side modified a file and the other side